+++ beacon-chain/rpc/beacon_chain_server.go
@@ -2,7 +2,9 @@ package rpc
 
 import (
 	"context"
+	"errors"
 	"sort"
+	"time"
 
 	ptypes "github.com/gogo/protobuf/types"
 	"github.com/prysmaticlabs/prysm/beacon-chain/blockchain"
@@ -12,6 +14,9 @@ import (
 	"github.com/prysmaticlabs/prysm/beacon-chain/db/filters"
 	"github.com/prysmaticlabs/prysm/beacon-chain/db/kv"
 	"github.com/prysmaticlabs/prysm/beacon-chain/operations"
+	"github.com/prysmaticlabs/prysm/beacon-chain/powchain"
+	pbp2p "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
+	pb "github.com/prysmaticlabs/prysm/proto/beacon/rpc/v1"
 	ethpb "github.com/prysmaticlabs/prysm/proto/eth/v1alpha1"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/pagination"
@@ -24,9 +29,15 @@ import (
 // providing RPC endpoints to access data relevant to the Ethereum 2.0 phase 0
 // beacon chain.
 type BeaconChainServer struct {
-	beaconDB    db.Database
-	headFetcher blockchain.HeadFetcher
-	pool        operations.Pool
+	beaconDB            db.Database
+	ctx                 context.Context
+	chainStartFetcher   powchain.ChainStartFetcher
+	headFetcher         blockchain.HeadFetcher
+	stateFeedListener   blockchain.ChainFeeds
+	pool                operations.Pool
+	incomingAttestation chan *ethpb.Attestation
+	canonicalStateChan  chan *pbp2p.BeaconState
+	chainStartChan      chan time.Time
 }
 
 // sortableAttestations implements the Sort interface to sort attestations
@@ -485,4 +496,40 @@ func (bs *BeaconChainServer) GetValidatorParticipation(
 		VotedEther:              attestedBalances,
 		EligibleEther:           totalBalances,
 	}, nil
+}
+
+// WaitForChainStart queries the logs of the Deposit Contract in order to verify the beacon chain
+// has started its runtime and validators begin their responsibilities. If it has not, it then
+// subscribes to an event stream triggered by the powchain service whenever the ChainStart log does
+// occur in the Deposit Contract on ETH 1.0.
+func (bs *BeaconChainServer) WaitForChainStart(req *ptypes.Empty, stream pb.BeaconService_WaitForChainStartServer) error {
+	head, err := bs.beaconDB.HeadState(context.Background())
+	if err != nil {
+		return err
+	}
+	if head != nil {
+		res := &pb.ChainStartResponse{
+			Started:     true,
+			GenesisTime: head.GenesisTime,
+		}
+		return stream.Send(res)
+	}
+
+	sub := bs.stateFeedListener.StateInitializedFeed().Subscribe(bs.chainStartChan)
+	defer sub.Unsubscribe()
+	for {
+		select {
+		case chainStartTime := <-bs.chainStartChan:
+			log.Info("Sending ChainStart log and genesis time to connected validator clients")
+			res := &pb.ChainStartResponse{
+				Started:     true,
+				GenesisTime: uint64(chainStartTime.Unix()),
+			}
+			return stream.Send(res)
+		case <-sub.Err():
+			return errors.New("subscriber closed, exiting goroutine")
+		case <-bs.ctx.Done():
+			return errors.New("rpc context closed, exiting goroutine")
+		}
+	}
 }
